Move and rename BaseDefinitionTest into BaseContainerTest
